move PersistAppliedConfig from public agent/phases/nodestart to cmd/agent/internal/daemon
