+++ cmd/agent/internal/daemon/persist_config.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-package nodestart
+package daemon
 
 import (
 	"context"
